requirements.txt
requests
PySimpleGUI
PyGithub
pyamiibo